🪲 [Fix]: Fix an issue where prereleases was not created during (#51)
## Description

- Reestablish functionality to run `Publish-PSModule` during PR. Used to
create prereleases of a module before it is published with a release
version.
- Revert: #37. This should have been reverted in the following PR,
seeing as it was for nightly runs/adhoc runs. The `CI.yml` template was
added just after which rendered this change unnecessary.
- It still remains to find a way to perserve the functionality, but not
need a pipeline to do it. I.e. a prereleaser "chat" bot that takes
commands from maintainers in the comments of the PR.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [x] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas