🪲 [Fix]: Fix docs generation (#133)
## Description

This pull request includes a small but important change to the
`Docs.yml` workflow configuration. The change modifies the path to the
`mkdocs.yml` configuration file to ensure the build process uses the
correct file.

*
[`.github/workflows/Docs.yml`](diffhunk://#diff-55dd8a146d52e27939b44e7e173a1c28589d78f9aaa2a3aa48edce6039534455L44-R44):
Changed the path of the `mkdocs.yml` configuration file from
`./mkdocs.yml` to `.github/mkdocs.yml` to correctly locate the
configuration during the build process.
* Put the actions documentation into an 'actions' folder on 'docs'.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [x] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas