🚀 [Feature]: Add functionality to configure the process with a settings file (#158)
## Description

This pull request includes significant changes to the CI workflow
configuration, primarily focusing on the introduction of a new
`Get-Settings` workflow and the modification of multiple job
dependencies to utilize this new workflow. The changes aim to streamline
the settings management and test suite configurations.

Key changes include:

### Workflow and Job Modifications:
*
[`.github/workflows/CI.yml`](diffhunk://#diff-3ab46ee209a127470fce3c2cf106b1a1dbadbb929a4b5b13656a4bc4ce19c0b8L35-R39):
Replaced the `Get-TestSuites` job with a new `Get-Settings` job, and
updated all subsequent job dependencies and inputs to use the outputs
from `Get-Settings`. This includes changes to job conditions, matrix
configurations, and input parameters.
[[1]](diffhunk://#diff-3ab46ee209a127470fce3c2cf106b1a1dbadbb929a4b5b13656a4bc4ce19c0b8L35-R39)
[[2]](diffhunk://#diff-3ab46ee209a127470fce3c2cf106b1a1dbadbb929a4b5b13656a4bc4ce19c0b8L62-R62)
[[3]](diffhunk://#diff-3ab46ee209a127470fce3c2cf106b1a1dbadbb929a4b5b13656a4bc4ce19c0b8L72-R95)
[[4]](diffhunk://#diff-3ab46ee209a127470fce3c2cf106b1a1dbadbb929a4b5b13656a4bc4ce19c0b8R104-R107)
[[5]](diffhunk://#diff-3ab46ee209a127470fce3c2cf106b1a1dbadbb929a4b5b13656a4bc4ce19c0b8R116-R199)

### New Workflow:
*
[`.github/workflows/Get-Settings.yml`](diffhunk://#diff-68f351d419a9b4db4797434891d5dfad44d83de65bfee7ba1c60badae369dbefR1-R338):
Added a new workflow, `Get-Settings`, which processes the settings file
and outputs necessary configurations for subsequent jobs. This includes
reading settings from various file formats (JSON, YAML, PSD1) and
determining test suites to run based on the settings.

### Additional Inputs:
*
[`.github/workflows/Get-CodeCoverage.yml`](diffhunk://#diff-261187083ecd8c20f585c08962cc876f48b4da8453718d673165a1a9cec9eecaR6-R15):
Added new inputs `StepSummary_Mode` and `CodeCoveragePercentTarget` to
control the GitHub step summary sections and set a target for code
coverage. These inputs are used in the `Get-CodeCoverage` job.
[[1]](diffhunk://#diff-261187083ecd8c20f585c08962cc876f48b4da8453718d673165a1a9cec9eecaR6-R15)
[[2]](diffhunk://#diff-261187083ecd8c20f585c08962cc876f48b4da8453718d673165a1a9cec9eecaR54-R55)

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [x] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas