🪲 [Fix]: Access to GH_TOKEN and Get-Settings for Name (#171)
## Description

This pull request includes updates to the workflow files to improve
clarity and functionality. The changes primarily focus on refining
conditional logic in PowerShell scripts and enhancing environment
variable handling in GitHub Actions.

### Workflow improvements:

*
[`.github/workflows/Get-Settings.yml`](diffhunk://#diff-68f351d419a9b4db4797434891d5dfad44d83de65bfee7ba1c60badae369dbefL134-R140):
Updated the conditional logic for determining the `$name` variable to
use explicit `if-elseif-else` statements, improving readability and
maintainability.

### Environment variable and input adjustments:

*
[`.github/workflows/workflow.yml`](diffhunk://#diff-126bf89616b7daa3d14ebc882ad18666aaf1c3dae888c4ba306a66ec80758bc1R289-L297):
Added the `GH_TOKEN` environment variable to the `Publish module` step
to ensure authentication is handled securely.
*
[`.github/workflows/workflow.yml`](diffhunk://#diff-126bf89616b7daa3d14ebc882ad18666aaf1c3dae888c4ba306a66ec80758bc1R305):
Reintroduced the `WorkingDirectory` input to the `Publish module` step
for better modularity and flexibility in specifying paths.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [x] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas